+++ src/views/PacketDetail/protocols/network/ARPDetails.ts
@@ -12,13 +12,13 @@ export function buildARPDetails(p: ARPPacket, offset: number): DetailsBuilderRes
       offset,
       length: p.raw.length,
       fields: [
-        {label: "Hardware Type", value: `${ARP_HARDWARE_TYPES[p.hardwareType]} (${p.hardwareType})`, offset: offset + 0, length: 2},
-        {label: "Protocol", value: `${ETHERTYPE_NAMES[p.protocol]} (${p.protocol})`, offset: offset + 2, length: 2},
-        {label: "Operation", value: `${ARP_OPERATIONS[p.operation]} (${p.operation})`, offset: offset + 6, length: 2},
-        {label: "Sender MAC", value: `${p.senderMAC}`, offset: offset + 8, length: 6},
-        {label: "Sender IP", value: `${p.senderIP}`, offset: offset + 14, length: 4},
-        {label: "Target MAC", value: `${p.targetMAC}`, offset: offset + 18, length: 6},
-        {label: "Target IP", value: `${p.targetIP}`, offset: offset + 24, length: 4}
+        {label: "Hardware Type", value: `${ARP_HARDWARE_TYPES[p.hardwareType]} (${p.hardwareType})`, offset:  0, length: 2},
+        {label: "Protocol", value: `${ETHERTYPE_NAMES[p.protocol]} (${p.protocol})`, offset:  2, length: 2},
+        {label: "Operation", value: `${ARP_OPERATIONS[p.operation]} (${p.operation})`, offset:  6, length: 2},
+        {label: "Sender MAC", value: `${p.senderMAC}`, offset:  8, length: 6},
+        {label: "Sender IP", value: `${p.senderIP}`, offset:  14, length: 4},
+        {label: "Target MAC", value: `${p.targetMAC}`, offset:  18, length: 6},
+        {label: "Target IP", value: `${p.targetIP}`, offset:  24, length: 4}
       ]
     }
   }