+++ src/views/PacketDetail/protocols/transport/ICMPv6Details.ts
@@ -18,8 +18,8 @@ export function buildICMPv6Details(p: ICMPv6Packet, offset: number): DetailsBuil
       offset,
       length: p.raw.length,
       fields: [
-        {label: "Type", value: `${typeName} (${p.icmpType})`, offset: offset + 0, length: 1},
-        {label: "Code", value: codeName ? `${codeName} (${p.code})` : `${p.code}`, offset: offset + 1, length: 1},
+        {label: "Type", value: `${typeName} (${p.icmpType})`, offset:  0, length: 1},
+        {label: "Code", value: codeName ? `${codeName} (${p.code})` : `${p.code}`, offset:  1, length: 1},
       ]
     }
   }